requirements.txt
streamlit==1.34.0
opencv-python-headless==4.9.0.80
pillow==10.3.0
